src/backend/requirements.txt
fastapi
uvicorn
sqlalchemy
psycopg2-binary
python-multipart
pydantic
python-docx
PyPDF2
python-multipart
python-jobspy
